fmt: forge fmt

+++ src/interfaces/IPublicAllocator.sol
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.21;
 
-import {IMetaMorpho, IMorpho, MarketAllocation,Id} from "../../lib/metamorpho/src/interfaces/IMetaMorpho.sol";
+import {IMetaMorpho, IMorpho, MarketAllocation, Id} from "../../lib/metamorpho/src/interfaces/IMetaMorpho.sol";
 
 struct FlowCaps {
     uint128 outflow;
     uint128 inflow;
-} 
+}
 
 struct FlowConfig {
     Id id;
@@ -14,16 +14,15 @@ struct FlowConfig {
     bool resetFlow;
 }
 
-
 /// @dev This interface is used for factorizing IPublicAllocatorStaticTyping and IPublicAllocator.
 /// @dev Consider using the IPublicAllocator interface instead of this one.
 interface IPublicAllocatorBase {
     function VAULT() external view returns (IMetaMorpho);
     function MORPHO() external view returns (IMorpho);
-    function flows(Id) external view returns (int);
+    function flows(Id) external view returns (int256);
 
     function reallocate(MarketAllocation[] calldata allocations) external payable;
-    function setFee(uint _fee) external;
+    function setFee(uint256 _fee) external;
     function transferFee(address feeRecipient) external;
     function setFlows(FlowConfig[] calldata configs) external;
 }
@@ -37,7 +36,8 @@ interface IPublicAllocatorStaticTyping is IPublicAllocatorBase {
 /// @title IPublicAllocator
 /// @author Morpho Labs
 /// @custom:contact security@morpho.org
-/// @dev Use this interface for PublicAllocator to have access to all the functions with the appropriate function signatures.
+/// @dev Use this interface for PublicAllocator to have access to all the functions with the appropriate function
+/// signatures.
 interface IPublicAllocator is IPublicAllocatorBase {
     function flowCaps(Id) external view returns (FlowCaps memory);
 }
